% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_temporal.R
\name{aggregate_temporal}
\alias{aggregate_temporal}
\title{Aggregate Temporal Data}
\usage{
aggregate_temporal(data, method = "sum", time_unit)
}
\arguments{
\item{data}{A data frame containing temporal data}

\item{method}{Aggregation method (e.g., "sum", "mean")}

\item{time_unit}{The target time unit for aggregation (e.g., "week", "month")}
}
\value{
Aggregated temporal data
}
\description{
Aggregates temporal data to a coarser time resolution
}
\examples{
\dontrun{
aggregated <- aggregate_temporal(data, "sum", "week")
}
}
