% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chap-package.R
\docType{package}
\name{chapr-package}
\alias{chapr}
\alias{chapr-package}
\title{chapr: Chap R SDK for Building Disease Forecasting Models}
\description{
Provides convenience functionality for developing Chap-compatible models and integrating existing models with the Chap (Climate Health Analytics Platform) system. Includes utilities for creating command line interfaces, model validation, spatio-temporal data transformations, and configuration schema management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dhis2-chap.github.io/chap_r_sdk/}
  \item \url{https://github.com/dhis2-chap/chap_r_sdk}
  \item Report bugs at \url{https://github.com/dhis2-chap/chap_r_sdk/issues}
}

}
\author{
\strong{Maintainer}: Knut Rand \email{knutdrand@gmail.com}

}
\keyword{internal}
