% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{detect_system_deps}
\alias{detect_system_deps}
\title{Detect system dependencies from renv packages}
\usage{
detect_system_deps(
  lockfile_path = "renv.lock",
  os = "ubuntu",
  os_release = "22.04"
)
}
\arguments{
\item{lockfile_path}{Path to renv.lock (default: "renv.lock")}

\item{os}{Operating system (default: "ubuntu")}

\item{os_release}{OS release version (default: "22.04")}
}
\value{
Character vector of apt package names
}
\description{
Parses renv.lock and queries package metadata to find required
system libraries (apt packages) using pak.
}
\examples{
\dontrun{
detect_system_deps()
detect_system_deps(os = "ubuntu", os_release = "24.04")
}
}
