% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{generate_dockerfile_content}
\alias{generate_dockerfile_content}
\title{Generate Dockerfile content}
\usage{
generate_dockerfile_content(
  base_image,
  system_deps,
  python_version,
  chapkit_spec
)
}
\arguments{
\item{base_image}{Base Docker image}

\item{system_deps}{Character vector of apt packages}

\item{python_version}{Python version}

\item{chapkit_spec}{Chapkit pip specification}
}
\value{
Character string with Dockerfile content
}
\description{
Internal function to generate the Dockerfile content string.
}
\keyword{internal}
