% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{make_info_parser}
\alias{make_info_parser}
\title{Create optparse parser for info subcommand}
\usage{
make_info_parser()
}
\value{
An OptionParser object
}
\description{
Create optparse parser for info subcommand
}
\keyword{internal}
