% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{make_predict_parser}
\alias{make_predict_parser}
\title{Create optparse parser for predict subcommand}
\usage{
make_predict_parser(
  default_config_path = "config.yml",
  default_model_path = "model.rds"
)
}
\arguments{
\item{default_config_path}{Default config file path}

\item{default_model_path}{Default model input path}
}
\value{
An OptionParser object
}
\description{
Create optparse parser for predict subcommand
}
\keyword{internal}
