% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.chap_config_schema}
\alias{print.chap_config_schema}
\title{Print Schema Summary}
\usage{
\method{print}{chap_config_schema}(x, ...)
}
\arguments{
\item{x}{Schema object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisible schema object
}
\description{
Prints a human-readable summary of a configuration schema.
}
