% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{schema_boolean}
\alias{schema_boolean}
\title{Define a Boolean Schema Property}
\usage{
schema_boolean(description = NULL, default = NULL)
}
\arguments{
\item{description}{Description of the parameter}

\item{default}{Default value if not provided}
}
\value{
A list representing a JSON Schema boolean property
}
\description{
Creates a JSON Schema property definition for a boolean value.
}
\examples{
schema_boolean(description = "Use cross-validation", default = TRUE)
}
