% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_temporal.R
\name{transform_spatiotemporal}
\alias{transform_spatiotemporal}
\title{Transform Spatio-Temporal Data}
\usage{
transform_spatiotemporal(data, transformation, ...)
}
\arguments{
\item{data}{A data frame or tibble containing spatio-temporal data}

\item{transformation}{The type of transformation to apply}

\item{...}{Additional arguments passed to the transformation function}
}
\value{
Transformed data
}
\description{
Provides common transformations for spatio-temporal data used in chap models
}
\examples{
\dontrun{
transformed <- transform_spatiotemporal(data, "normalize")
}
}
