% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chap-package.R
\docType{package}
\name{chap.r.sdk-package}
\alias{chap.r.sdk}
\alias{chap.r.sdk-package}
\title{chap.r.sdk: Chap SDK for R}
\description{
Provides convenience functionality for developing chap-compatible models and integrating existing models with chap. Includes utilities for creating command line interfaces, model validation, spatio-temporal data transformations, and configuration schema management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/knutdrand/chap_r_sdk}
  \item Report bugs at \url{https://github.com/knutdrand/chap_r_sdk/issues}
}

}
\author{
\strong{Maintainer}: Knut Rand \email{knutdrand@gmail.com}

}
\keyword{internal}
